# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_yourusername_noshirvani-academy_backend_internal_domain.BlogPost:
    properties:
      author_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      published:
        type: boolean
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_yourusername_noshirvani-academy_backend_internal_domain.Student:
    properties:
      approval_date:
        type: string
      birth_date:
        type: string
      city:
        type: string
      created_at:
        type: string
      dynamic_fields:
        additionalProperties: true
        type: object
      first_name:
        type: string
      id:
        type: string
      is_approved:
        type: boolean
      jalali_birth_date:
        type: string
      last_name:
        type: string
      major:
        type: string
      profile_photo:
        type: string
      school:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.User'
      user_id:
        type: string
    type: object
...
  github_com_yourusername_noshirvani-academy_backend_internal_domain.User:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      role:
        type: string
      student:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Student'
      updated_at:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handler.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        example: 3600
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.User'
    type: object
  internal_handler.BlogInput:
    properties:
      author_id:
        type: string
      content:
        type: string
      published:
        example: false
        type: boolean
      slug:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handler.OTPResponse:
    properties:
      message:
        example: otp sent
        type: string
      otp:
        example: "123456"
        type: string
    type: object
  internal_handler.RefreshTokenInput:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_handler.RequestOTPInput:
    properties:
      phone:
        example: "+989123456789"
        type: string
    required:
    - phone
    type: object
  internal_handler.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        example: 3600
        type: integer
    type: object
...
  internal_handler.VerifyOTPInput:
    properties:
      code:
        example: "123456"
        type: string
      phone:
        example: "+989123456789"
        type: string
    required:
    - code
    - phone
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@noshirvaniacademy.com
    name: تیم توسعه اپلیکیشن
    url: https://github.com/Offline-Devs
  description: |-
    API برای سیستم مدیریت آکادمی نوشیروانی
    شامل خدماتی برای احراز هویت، مدیریت دانشجویان، آزمون‌ها، اشتباهات و مقالات بلاگ
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: نوشیروانی آکادمی API
  version: "1.0"
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: با استفاده از توکن تازه‌سازی، توکن دسترسی جدید ایجاد می‌کند
      parameters:
      - description: توکن تازه‌سازی
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: توکن جدید با موفقیت ایجاد شد
          schema:
            $ref: '#/definitions/internal_handler.TokenResponse'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: توکن تازه‌سازی نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: تازه‌سازی توکن دسترسی
      tags:
      - احراز هویت
  /auth/request-otp:
    post:
      consumes:
      - application/json
      description: کد OTP را برای شماره تلفن کاربر ارسال می‌کند
      parameters:
      - description: شماره تلفن
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RequestOTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: کد OTP با موفقیت ارسال شد
          schema:
            $ref: '#/definitions/internal_handler.OTPResponse'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: درخواست کد OTP
      tags:
      - احراز هویت
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: کد OTP را تأیید می‌کند و توکن‌های JWT برای کاربر ایجاد می‌کند
      parameters:
      - description: شماره تلفن و کد OTP
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.VerifyOTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: احراز هویت موفق
          schema:
            $ref: '#/definitions/internal_handler.AuthResponse'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: کد OTP نامعتبر یا منقضی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: کاربر غیرفعال است
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: تأیید کد OTP و ورود کاربر
      tags:
      - احراز هویت
  /blog:
    get:
      description: تمام نوشته‌های منتشر‌شده را دریافت می‌کند (بدون نیاز به احراز هویت)
      produces:
      - application/json
      responses:
        "200":
          description: لیست نوشته‌های منتشر‌شده
          schema:
            items:
              $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.BlogPost'
            type: array
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: دریافت لیست نوشته‌های منتشر شده
      tags:
      - بلاگ
    post:
      consumes:
      - application/json
      description: یک نوشته بلاگ جدید را ایجاد می‌کند (فقط برای مدیران)
      parameters:
      - description: اطلاعات نوشته
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.BlogInput'
      produces:
      - application/json
      responses:
        "201":
          description: نوشته با موفقیت ایجاد شد
          schema:
            $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.BlogPost'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ایجاد نوشته بلاگ جدید
      tags:
      - بلاگ
  /blog/{id}:
    delete:
      description: یک نوشته بلاگ را حذف می‌کند (فقط برای مدیران)
      parameters:
      - description: شناسه نوشته
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: نوشته با موفقیت حذف شد
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: حذف نوشته بلاگ
      tags:
      - بلاگ
    put:
      consumes:
      - application/json
      description: اطلاعات یک نوشته بلاگ را بروزرسانی می‌کند (فقط برای مدیران)
      parameters:
      - description: شناسه نوشته
        in: path
        name: id
        required: true
        type: string
      - description: اطلاعات جدید نوشته
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.BlogInput'
      produces:
      - application/json
      responses:
        "200":
          description: نوشته با موفقیت بروزرسانی شد
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: بروزرسانی نوشته بلاگ
      tags:
      - بلاگ
  /blog/{id}/publish:
    put:
      description: یک نوشته بلاگ را برای عموم قابل دسترس می‌کند (فقط برای مدیران)
      parameters:
      - description: شناسه نوشته
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: نوشته با موفقیت منتشر شد
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: انتشار نوشته بلاگ
      tags:
      - بلاگ
  /blog/{slug}:
    get:
      description: یک نوشته منتشر شده را با استفاده از slug آن دریافت می‌کند (بدون
        نیاز به احراز هویت)
      parameters:
      - description: نشانی URL نوشته
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: نوشته بلاگ
          schema:
            $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.BlogPost'
        "404":
          description: نوشته یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: دریافت نوشته بلاگ با slug
      tags:
      - بلاگ
...
securityDefinitions:
  BearerAuth:
    description: '"توکن JWT - فرمت: Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"